requirements.txt
allure-pytest==2.9.45
pytest==6.2.5
selenium==3.141.0
webdriver-manager==3.5.2
requests==2.26.0
pytest-xdist==2.4.0
pytest-rerunfailures==10.2
pymongo==3.12.1
PyMySQL==1.0.2
pyhamcrest==2.0.2
imap-tools==0.49.1
paramiko==2.8.0
python-dotenv==0.19.1
